backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
bcrypt==3.2.0
passlib[bcrypt]
python-jose[cryptography]
python-multipart
openai
email-validator